SPT-1514: Refactor code into `identity-reuse` directory
Create new `identity-reuse` directory for the code that contributes
to the decision making around whether an identity can be reused or not.
We want to keep this code separate from the handlers code as we could
end up creating more handlers for future endpoints that will use this
code and allow more flexibility.
