+++ src/identity-reuse/verifiable-credential-jwt.ts
